Move default or no default logic into native.argument
Previously, it was done at a use site, but that meant other use
sites don't get the right logic.  Pushing it in makes sure everyone
gets it.

I also fixed one case of confusion where defn() was used to define a decl().
If you want to define a declaration with no defaults, say no_default().decl()
which is more direct and will give us code reviewers a clue if you should
have pushed this logic in.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 79f8c856bbe0570b9b52a01e0c61571e1633d946
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49489